pyproject.toml
[project]
name = "atomdoc"
version = "0.1.0"
description = "Local-first document models with semantic atomicity"
readme = "README.md"
authors = [
    { name = "MH", email = "mhalle@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "python-ulid>=3.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=6.0",
    "ruff>=0.9",
    "mypy>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/atomdoc"]

[tool.mypy]
plugins = ["pydantic.mypy"]
warn_return_any = true
warn_unused_configs = true
check_untyped_defs = true

[tool.ruff]
target-version = "py312"

[tool.pytest.ini_options]
testpaths = ["tests"]
